Deduplicate events via uid
add a uid to events of the form: date, hash of event type + name + hebrewdate, and the domain.

This conforms to the ical uid spec, and most clients should deduplicate events with the same uid from different calendars. the use case is if one enters the same birthday in different family calendars, the event should only show once, es it is realy the same event

untested